requirements.txt
django==1.11
django-allauth